+++ src/components/visualEditor/ActionForms/matchActionForm.tsx
@@ -14,7 +14,7 @@ import {isState, MatchActionProps, ReactFlowContextProps} from "../../../types.t
 import MatchCase from "../../../classes/MatchCase.tsx";
 
 
-let idCounter = 0
+
 
 function CaseForm(props: {
     action: Action | undefined;
@@ -25,10 +25,8 @@ function CaseForm(props: {
 }) {
 
 
-    const instanceId = useRef(++idCounter).current;
-    let formCount = 0
+    const formId = useRef(`form-${Math.random().toString(36).substring(2, 15)}`).current;
 
-    const formId = `${instanceId}-${formCount++}`
 
     useEffect(() => {
         console.log(`FORM ID : ${formId}`)
@@ -157,13 +155,13 @@ function CaseForm(props: {
             case ActionType.ASSIGN:
                 return <AssignActionForm action={props.action ? props.action : assignAction[0]} setActions={props.action ? props.setActions : setAssignAction} noCategorySelect={true} dontAddToState={true} />;
             case ActionType.RAISE_EVENT:
-                return <RaiseEventActionForm action={props.action ? props.action : raiseEventAction[0]} setActions={props.action ? props.setActions : setRaiseEventAction} noCategorySelect={true} />;
+                return <RaiseEventActionForm action={props.action ? props.action : raiseEventAction[0]} setActions={props.action ? props.setActions : setRaiseEventAction} noCategorySelect={true}  dontAddToEdge={true}/>;
             case ActionType.TIMEOUT:
-                return <TimeoutActionForm action={props.action ? props.action : timeoutAction[0]} setActions={props.action ? props.setActions : setTimeoutAction} noCategorySelect={true} />;
+                return <TimeoutActionForm action={props.action ? props.action : timeoutAction[0]} setActions={props.action ? props.setActions : setTimeoutAction} noCategorySelect={true} dontAddToState={true} />;
             case ActionType.TIMEOUT_RESET:
-                return <TimeoutResetActionForm action={props.action ? props.action : timeOutResetAction[0]} setActions={props.action ? props.setActions : setTimeoutResetAction} noCategorySelect={true} />;
+                return <TimeoutResetActionForm action={props.action ? props.action : timeOutResetAction[0]} setActions={props.action ? props.setActions : setTimeoutResetAction} noCategorySelect={true} dontAddToState={true} />;
             case ActionType.MATCH:
-                return <MatchActionForm action={props.action ? props.action : matchAction[0]} setActions={props.action ? props.setActions : setMatchAction}  dontAddToState={true}/>;
+                return <MatchActionForm action={props.action ? props.action : matchAction[0]} setActions={props.action ? props.setActions : setMatchAction} noCategorySelect={true} dontAddToState={true}/>;
             default:
                 return null;
         }
@@ -244,15 +242,14 @@ export default function MatchActionForm(props: {
 
 }) {
 
-    const instanceId = useRef(++idCounter).current;
-    let formCount = 0
 
-    const formId = `${instanceId}-${formCount++}`
+    const formId = useRef(`form-${Math.random().toString(36).substring(2, 15)}`).current;
+
     const submitButtonText = () => props.action ? "Save Changes" : "Create Action"
 
 
     const context = useContext(ReactFlowContext) as ReactFlowContextProps;
-    const { selectedNode , actionService, stateOrStateMachineService} = context;
+    const { selectedNode , actionService, stateOrStateMachineService, selectedEdge} = context;
 
     const [expressionInput, setExpressionInput] = useState<string>("");
     const [caseActions, setCaseActions] = useState<Action[]>([]);
@@ -268,6 +265,10 @@ export default function MatchActionForm(props: {
         setShow(true);
     };
 
+    useEffect(() => {
+        console.log(`FORM ID : ${formId}`)
+        console.log(selectedEdge)
+    }, []);
     const handleClose = () => {
         setShow(false);
     };
@@ -300,7 +301,8 @@ export default function MatchActionForm(props: {
 
 
     useEffect(() => {
-            if(!selectedNode){
+
+            if(!selectedNode && ! selectedEdge){
                 return
             }
             if(props.action && props.action.type === ActionType.MATCH) {
@@ -309,11 +311,12 @@ export default function MatchActionForm(props: {
                 const actions = matchActionProps.cases.map((mc) => mc.action)
                 setCaseActions(actions)
 
-                const actionCategory = actionService.getActionCategory(props.action,selectedNode.data)
-                if(actionCategory) {
-                    setSelectedActionCategory(actionCategory)
+                if(selectedNode){
+                    const actionCategory = actionService.getActionCategory(props.action,selectedNode.data)
+                    if(actionCategory) {
+                        setSelectedActionCategory(actionCategory)
+                    }
                 }
-
 
             }
     }, [props.action, selectedActionCategory]);
@@ -334,7 +337,7 @@ export default function MatchActionForm(props: {
 
     const onDeleteButtonPress = () => {
 
-        if(!selectedNode){
+        if(!selectedNode && !selectedEdge){
             return;
         }
 
@@ -342,7 +345,15 @@ export default function MatchActionForm(props: {
             return
         }
 
-        stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data)
+        if(selectedNode){
+            stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data)
+        }
+
+        if(selectedEdge?.data){
+            selectedEdge.data.transition.removeAction(props.action)
+        }
+
+
         props.setActions((prevActions) => prevActions.filter((a) => a !== props.action))
 
     }
@@ -353,7 +364,12 @@ export default function MatchActionForm(props: {
         event.preventDefault();
         event.stopPropagation();
 
-        if (!selectedNode || caseActions.length < 1) {
+        console.log("hi")
+        console.log(selectedEdge)
+
+        //*
+
+        if ((!selectedNode && !selectedEdge) || caseActions.length < 1) {
             return;
         }
 
@@ -375,46 +391,98 @@ export default function MatchActionForm(props: {
 
         let updatedAction: Action
 
-        if (props.action && props.action.type === ActionType.MATCH) {
+        if(selectedEdge){
 
-            const oldCategory = actionService.getActionCategory(props.action, selectedNode.data);
+            console.log("Mate")
 
-            // Check if category has changed and update the action in the state node if needed
-            if (oldCategory !== selectedActionCategory as ActionCategory) {
-                if(!props.dontAddToState){
-                    stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
-                    stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+            if (props.action && props.action.type === ActionType.MATCH) {
+                updatedAction = props.action;
+                updatedAction.properties = matchActionProps;
+                onActionSubmit(updatedAction);
+
+                console.log("EXISTING")
+
+                actionService.deregisterAction(updatedAction);
+                actionService.registerAction(updatedAction);
+
+                if(selectedEdge.data && !props.dontAddToState){
+                    selectedEdge.data.transition.removeAction(updatedAction)
+                    selectedEdge.data.transition.addAction(updatedAction)
                 }
+
+                if (props.onSubmit) {
+                    props.onSubmit();
+                }
+                return
+
+
+            }
+            else {
+                console.log("NEW")
+                updatedAction = new Action("", ActionType.MATCH);
+                updatedAction.properties = matchActionProps;
+
+
+                actionService.deregisterAction(updatedAction);
+                actionService.registerAction(updatedAction);
+
+
+                onActionSubmit(updatedAction);
+
+
+
+                if(selectedEdge.data && !props.dontAddToState){
+                    selectedEdge.data.transition.addAction(updatedAction)
+                }
+                if (props.onSubmit) {
+                    props.onSubmit();
+                }
+
+            }
+            return
+        }
+
+        if(selectedNode) {
+            if (props.action && props.action.type === ActionType.MATCH) {
+
+                const oldCategory = actionService.getActionCategory(props.action, selectedNode.data);
+
+                // Check if category has changed and update the action in the state node if needed
+                if (oldCategory !== selectedActionCategory as ActionCategory) {
+                    if (!props.dontAddToState) {
+                        stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
+                        stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+                    }
+                }
+
+                updatedAction = props.action;
+                updatedAction.properties = matchActionProps;
+                onActionSubmit(updatedAction);
+            } else {
+                updatedAction = new Action("", ActionType.MATCH);
+                updatedAction.properties = matchActionProps;
+                if (!props.dontAddToState) {
+                    stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
+                }
+                onActionSubmit(updatedAction);
             }
 
-            updatedAction = props.action;
-            updatedAction.properties = matchActionProps;
-            onActionSubmit(updatedAction);
-        } else {
-            updatedAction = new Action("", ActionType.MATCH);
-            updatedAction.properties = matchActionProps;
-            if(!props.dontAddToState){
-                stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
+            if (props.onSubmit) {
+                props.onSubmit();
             }
-            onActionSubmit(updatedAction);
+
+            actionService.deregisterAction(updatedAction);
+            actionService.registerAction(updatedAction);
+
+
+            if (isState(selectedNode.data)) {
+                const sm = selectedNode.data
+                sm.state.entry.forEach(entry => {
+                    console.log(entry.properties)
+                })
+            }
         }
 
-        if (props.onSubmit) {
-            props.onSubmit();
-        }
-
-        actionService.deregisterAction(updatedAction);
-        actionService.registerAction(updatedAction);
-
-
-        if(isState(selectedNode.data)){
-            const sm = selectedNode.data
-            sm.state.entry.forEach(entry => {
-                console.log(entry.properties)
-            })
-        }
-
-        // Add logic for MatchAction submission if needed
     };
 
     return (
