+++ src/components/visualEditor/ActionForms/timeoutActionForm.tsx
@@ -11,10 +11,10 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
     setActions: Dispatch<SetStateAction<Action[]>>,
     onSubmit?: () => void,
     noCategorySelect? :boolean,
-    dontShowDeleteButton? :boolean}) {
+    dontShowDeleteButton? :boolean, dontAddToState?: boolean}) {
 
     const context = useContext(ReactFlowContext) as ReactFlowContextProps;
-    const {selectedNode, actionService, stateOrStateMachineService, eventService} = context
+    const {selectedNode, actionService, stateOrStateMachineService, eventService, selectedEdge} = context
 
     const oldName = props.action ? props.action.name : undefined
     const formID = "timeoutForm"
@@ -122,7 +122,7 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
     }, [actionNameInput,timeoutExpressionInput,timeoutAction]);
 
     useEffect(() => {
-        if(!selectedNode){
+        if(!selectedNode && !selectedEdge){
             return
         }
         if(props.action && props.action.type === ActionType.TIMEOUT){
@@ -133,9 +133,11 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
             setActionNameInput(timeoutActionProps.name)
             setTimeoutExpressionInput(timeoutActionProps.delay)
 
-            const category = actionService.getActionCategory(props.action,selectedNode.data)
-            if(category){
-                setSelectedActionCategory(category)
+            if(selectedNode){
+                const category = actionService.getActionCategory(props.action,selectedNode.data)
+                if(category){
+                    setSelectedActionCategory(category)
+                }
             }
 
             console.log(!!timeoutAction)
@@ -150,7 +152,7 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
 
     const onDeleteButtonPress = () => {
 
-        if(!selectedNode){
+        if(!selectedNode && !selectedEdge){
             return;
         }
 
@@ -161,7 +163,14 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
         const raiseEventProps = (props.action.properties as TimeoutActionProps).action.properties as RaiseEventActionProps;
 
         eventService.unregisterEvent(raiseEventProps.event.name)
-        stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data)
+        if(selectedNode){
+            stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data)
+        }
+
+        if(selectedEdge?.data){
+            selectedEdge.data.transition.removeAction(props.action)
+        }
+
         props.setActions((prevActions) => prevActions.filter((a) => a !== props.action))
 
     }
@@ -169,7 +178,7 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
     const onSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         event.stopPropagation();
-        if(!selectedNode){
+        if(!selectedNode && !selectedEdge){
             return
         }
 
@@ -186,40 +195,88 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
 
         let updatedAction: Action;
 
-        // Handle updating an existing action
-        if (props.action) {
 
-            const oldCategory = actionService.getActionCategory(props.action, selectedNode.data);
+        if(selectedEdge){
 
-            // Check if category has changed and update the action in the state node if needed
-            if (oldCategory !== selectedActionCategory as ActionCategory) {
-                stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
-                stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+            if(props.action){
+                updatedAction = props.action;
+                updatedAction.properties = timeoutActionProps;
+                onActionSubmit(updatedAction);
+
+
+
+                actionService.deregisterAction(updatedAction);
+                actionService.registerAction(updatedAction);
+
+                if(selectedEdge.data && !props.dontAddToState){
+                    selectedEdge.data.transition.removeAction(updatedAction)
+                    selectedEdge.data.transition.addAction(updatedAction)
+                }
+
+                if (props.onSubmit) {
+                    props.onSubmit();
+                }
+                return
+            }
+            else {
+                updatedAction = new Action(timeoutActionProps.name, ActionType.TIMEOUT);
+                updatedAction.properties = timeoutActionProps;
+                onActionSubmit(updatedAction);
+
+                if(selectedEdge.data && !props.dontAddToState){
+                    selectedEdge.data.transition.addAction(updatedAction)
+                }
+
+                actionService.registerAction(updatedAction);
+
+                if (props.onSubmit) {
+                    props.onSubmit();
+                }
+            }
+            return
+
+        }
+
+
+        if(selectedNode) {
+
+            // Handle updating an existing action
+            if (props.action) {
+
+                const oldCategory = actionService.getActionCategory(props.action, selectedNode.data);
+
+                // Check if category has changed and update the action in the state node if needed
+                if (oldCategory !== selectedActionCategory as ActionCategory && !props.dontAddToState) {
+                    stateOrStateMachineService.removeActionFromState(props.action, selectedNode.data);
+                    stateOrStateMachineService.addActionToState(selectedNode.data, props.action, selectedActionCategory as ActionCategory);
+                }
+
+                updatedAction = props.action;
+                updatedAction.properties = timeoutActionProps;
+                onActionSubmit(updatedAction);
+            } else {
+                updatedAction = new Action(timeoutActionProps.name, ActionType.TIMEOUT);
+                updatedAction.properties = timeoutActionProps;
+                if(!props.dontAddToState){
+                    stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
+                }
+                onActionSubmit(updatedAction);
             }
 
-            updatedAction = props.action;
-            updatedAction.properties = timeoutActionProps;
-            onActionSubmit(updatedAction);
-        } else {
-            updatedAction = new Action(timeoutActionProps.name, ActionType.TIMEOUT);
-            updatedAction.properties = timeoutActionProps;
-            stateOrStateMachineService.addActionToState(selectedNode.data, updatedAction, selectedActionCategory as ActionCategory);
-            onActionSubmit(updatedAction);
-        }
+            if (props.onSubmit) {
+                props.onSubmit();
+            }
 
-        if (props.onSubmit) {
-            props.onSubmit();
-        }
-
-        actionService.deregisterAction(updatedAction);
-        actionService.registerAction(updatedAction);
+            actionService.deregisterAction(updatedAction);
+            actionService.registerAction(updatedAction);
 
 
-        if(isState(selectedNode.data)){
-            const sm = selectedNode.data
-            sm.state.entry.forEach(entry => {
-                console.log(entry.properties)
-            })
+            if (isState(selectedNode.data)) {
+                const sm = selectedNode.data
+                sm.state.entry.forEach(entry => {
+                    console.log(entry.properties)
+                })
+            }
         }
 
 
@@ -288,6 +345,7 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
                        action={timeoutAction[0]}
                        onSubmit={onActionFormSubmit}
                        singleAction={true}
+                       dontAddToEdge={true}
                        noCategorySelect={true}
                    />
                ) : (
@@ -296,6 +354,7 @@ export default function TimeoutActionForm(props: {action: Action | undefined,
                        action={undefined}
                        onSubmit={onActionFormSubmit}
                        singleAction={true}
+                       dontAddToEdge={true}
                        noCategorySelect={true}
                    />
                )}
